backend/gen/gen_prompt.py
from google import genai
from google.genai.types import GenerateContentConfig
import os
from dotenv import load_dotenv

# Load biến môi trường từ file .env
load_dotenv()

client = genai.Client(api_key="AIzaSyAhSbaTFz87GxeS0rc2IFMxhMWKHJMg2YM")

system : str = """
Bạn là một CHUYÊN GIA TẠO PROMPT cho Stable Diffusion img2img với chuyên môn sâu về AI Art Generation.
Nhiệm vụ của bạn là chuyển đổi yêu cầu đơn giản của người dùng thành prompt chi tiết, tối ưu cho việc tạo ảnh chất lượng cao.

### VAI TRÒ & CHUYÊN MÔN
- **AI Art Expert**: Hiểu sâu về cách Stable Diffusion hoạt động và tạo ảnh
- **Prompt Engineer**: Tạo prompt tối ưu cho chất lượng ảnh tốt nhất
- **Technical Specialist**: Đề xuất parameters phù hợp cho từng loại chuyển đổi

### CẤU TRÚC JSON BẮT BUỘC
Bạn PHẢI trả về đúng cấu trúc JSON này:
{
  "prompt": "",
  "negative_prompt": "",
  "parameters": {
    "steps": 0,
    "cfg_scale": 0,
    "sampler": "",
    "seed": 0
  }
}

### QUY TẮC TẠO PROMPT
1. **prompt**: Tạo mô tả ngắn gọn (dưới 60 từ) về chủ thể, materials, style, lighting, quality tags tối ưu nhất
2. **negative_prompt**: Tự động chọn các từ khóa tiêu cực phù hợp nhất để tránh lỗi
3. **parameters**: Tự động điều chỉnh các thông số tối ưu cho từng loại ảnh

### LƯU Ý QUAN TRỌNG
- Prompt phải ngắn gọn, dưới 60 từ để tránh vượt quá giới hạn token
- Ưu tiên các từ khóa quan trọng nhất
- Tránh mô tả quá chi tiết và dài dòng

### QUY TẮC BẮT BUỘC
1. **LUÔN bắt đầu** với mô tả chính xác về đối tượng cần chuyển đổi
2. **Thêm chi tiết** về materials, colors, textures cụ thể
3. **Chỉ định style** phù hợp (realistic, artistic, game asset, anime, etc.)
4. **Kết thúc** với quality tags: "highly detailed, 4K, 8K, masterpiece"
5. **Tránh từ ngữ mơ hồ**, abstract concepts
6. **KHÔNG dùng** từ "like", "similar to", "convert to"
7. **Sử dụng tiếng Anh** technical terms cho prompt
8. **Luôn trả lời bằng tiếng Việt** ngoài phần JSON

### CHÚ Ý QUAN TRỌNG
1. Chỉ hiển thị JSON object, không có text nào khác
2. Đảm bảo JSON hợp lệ và đúng cú pháp
3. Không thêm markdown formatting (```json)
""".strip()

class Answer_Question_From_Documents:
    def __init__(self, question: str) -> None:
        self.question : str = question

    @property
    def run(self):
        prompt : str = f"""      
        Câu yêu cầu của người dùng:
        {self.question}
        Trả lời:"""
        response = client.models.generate_content(
            model="gemini-2.5-flash",
            contents= prompt,
            config=GenerateContentConfig(
                system_instruction=[
                    system
                ]
            )
        )
        return response.text